# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: FLibrary Web API
  version: 0.1.0
  description: |
    API веб-каталогизатора поверх коллекции FLibrary.

    Коллекционная БД (`collection.db`) читается только на чтение — её схема принадлежит
    FLibrary. Всё своё (пользователи, их данные, поисковые индексы) живёт в `app.db`.
    Бинарные ручки (обложки, файлы книг) проксируются во внутренний C++-сервер `opds`.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Вход по логину и паролю
      description: При успехе ставит httpOnly-куку сессии. Регистрации нет — пользователей создаёт администратор.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [login, password]
              properties:
                login: { type: string, minLength: 1, maxLength: 64 }
                password: { type: string, minLength: 1, maxLength: 256 }
      responses:
        '200':
          description: Успешный вход
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /me:
    get:
      tags: [auth]
      summary: Текущий пользователь
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /books/{bookId}:
    get:
      tags: [books]
      summary: Карточка книги
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /books/{bookId}/cover:
    get:
      tags: [books]
      summary: Обложка книги
      description: Отдаётся из дискового кэша; при первом запросе извлекается из архива внутренним C++-сервером.
      parameters:
        - $ref: '#/components/parameters/BookId'
        - name: size
          in: query
          schema:
            enum: [thumb, full]
            default: thumb
      responses:
        '200':
          description: Изображение обложки
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
          content:
            image/webp:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
        '304': { description: Не изменилось }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /books/{bookId}/content:
    get:
      tags: [books]
      summary: Файл книги
      description: Стримится из архива внутренним C++-сервером. Форматы кроме исходного требуют настроенных конвертеров.
      parameters:
        - $ref: '#/components/parameters/BookId'
        - name: format
          in: query
          schema:
            enum: [original, fb2, zip, epub, mobi]
            default: original
      responses:
        '200':
          description: Файл книги
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/BadGateway' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: flw_session
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Слишком много попыток
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    BadGateway:
      description: Внутренний content-service недоступен
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      description: Ошибка в формате RFC 9457.
      type: object
      required: [title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
    User:
      type: object
      required: [login, displayName]
      properties:
        login: { type: string }
        displayName: { type: string }
...
    BookListItem:
      type: object
      required: [bookId, title, authors, hasCover]
      properties:
        bookId: { type: integer }
        title: { type: string }
        authors:
          type: array
          items: { $ref: '#/components/schemas/AuthorRef' }
        series: { $ref: '#/components/schemas/SeriesRef' }
        seqNumber: { type: [integer, 'null'] }
        year: { type: [integer, 'null'] }
        lang: { type: [string, 'null'] }
        ext: { type: [string, 'null'] }
        size: { type: [integer, 'null'] }
        libRate: { type: [integer, 'null'] }
        userRate: { type: [integer, 'null'] }
        genres:
          type: array
          items: { $ref: '#/components/schemas/GenreRef' }
        hasCover: { type: boolean }
        favorite: { type: boolean }
        read: { type: boolean }
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/BookListItem'
        - type: object
          properties:
            annotation: { type: [string, 'null'] }
            keywords:
              type: array
              items: { type: string }
            archive: { type: [string, 'null'] }
            fileName: { type: [string, 'null'] }
            updateDate: { type: [string, 'null'] }
            formats:
              description: Форматы, в которых книгу можно скачать.
              type: array
              items: { type: string }
            seriesBooks:
              description: Соседние книги серии для навигации.
              type: array
              items: { $ref: '#/components/schemas/BookListItem' }
    AuthorRef:
      type: object
      required: [authorId, name]
      properties:
        authorId: { type: integer }
        name: { type: string }
        lastName: { type: [string, 'null'] }
        firstName: { type: [string, 'null'] }
        middleName: { type: [string, 'null'] }
    SeriesRef:
      type: [object, 'null']
      required: [seriesId, title]
      properties:
        seriesId: { type: integer }
        title: { type: string }
    GenreRef:
      type: object
      required: [code, title]
      properties:
        code: { type: string }
        title: { type: string }
...